requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0
psycopg2-binary==2.8.3
gunicorn==19.9.0
numpy==1.18.1
scipy==1.4.1
pandas==0.25.3
requests==2.7.0
lxml==4.5.0
joblib==0.14.1
scikit-learn==0.21.3
category-encoders==2.1.0
nose2==0.9.2 
